Support XMPP notifications
On the Admin page, when adding a new server, one can provide an XMPP
address which should be notified when the server goes down.  As soon as
the server is added, an XMPP invite is sent requesting permission to
communicate with that address.

When the server goes down, a short message is sent to the XMPP address
with details.